send question refactor
