+++ controllers/biddings/questions/index.js
@@ -1,6 +1,5 @@
-const logger = require('winston-namespace')('bidding:questions')
-const {Bidding, User} = require('../../../models')
-const {token} = require('../../auth')
+const {Bidding} = require('../../../models')
+const {getUserIdByToken} = require('../../utils')
 
 /**
  * Updates a bidding with the data coming from the body of the request.
@@ -38,27 +37,6 @@ function update (req, res, next) {
       next(err)
     })
   next()
-}
-
-// TODO: remove this function (duplicated). It should be imported from wherever the original function is.
-function getUserIdByToken (tkn) {
-  return token.getData(tkn)
-    .then(tokenData => {
-      return tokenData.email
-    }).then(email => {
-      return User.findOne({email: email})
-        .then(user => {
-          if (!user) {
-            const err = new Error(`Unexpected: User with email '${email}' not found`)
-            err.status = 404
-            throw err
-          }
-          return user._id
-        })
-    })
-    .catch(err => {
-      throw err
-    })
 }
 
 module.exports = {
